add support for sensor readings

+++ src/Application/DTOs/Sensor/CreateSensor.cs
@@ -8,8 +8,8 @@ public class CreateSensor
     public required string Name { get; set; }
     public required string Description { get; set; }
     public required string MqttTopic { get; set; }
-    public required int MinReading { get; set; }
-    public required int MaxReading { get; set; }
+    public required double MinReading { get; set; }
+    public required double MaxReading { get; set; }
     public required string Measurement { get; set; }
 
     public static explicit operator Sensor(CreateSensor createSensor)
